Adds 'typesVersions' support to NodeJS definitions
