+++ types/node/ts3.1/node-tests.ts
@@ -1,0 +1,2 @@
+// tslint:disable-next-line:no-bad-reference
+import "../node-tests";
