+++ types/node/ts3.1/util.d.ts
@@ -1,0 +1,10 @@
+// tslint:disable-next-line:no-bad-reference
+/// <reference path="../util.d.ts" />
+declare module "util" {
+    namespace inspect {
+        const custom: unique symbol;
+    }
+    namespace promisify {
+        const custom: unique symbol;
+    }
+}
